push_swap ongoing / basic operations

+++ libft/ft_atoi.c
@@ -6,13 +6,13 @@
 /*   By: fmoulin <fmoulin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/30 13:46:47 by fmoulin           #+#    #+#             */
-/*   Updated: 2025/05/06 19:36:20 by fmoulin          ###   ########.fr       */
+/*   Updated: 2025/07/01 15:23:58 by fmoulin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *nptr)
+size_t	ft_atoi(const char *nptr)
 {
 	size_t	i;
 	size_t	sign;
